Move StripedLocks to commons-lang module
